task-8

+++ specialities/services.py
@@ -2,8 +2,10 @@ import models
 
 
 def get_speciality(target_id):
-    speciality = models.Speciality.objects.get(id=target_id)
-    return speciality
+    if models.Speciality.objects.filter(id=target_id).exist():
+        speciality = models.Speciality.objects.get(id=target_id)
+        return speciality
+    return None
 
 
 def add_speciality(name: str, code: str, level: str, form: str):
@@ -14,9 +16,10 @@ def add_speciality(name: str, code: str, level: str, form: str):
 
 
 def delete_speciality(target_id: int):
-    speciality = models.Speciality.objects.get(id=target_id)
-    speciality.delete()
-    return speciality is None
+    if models.Speciality.objects.filter(id=target_id).exist():
+        speciality = models.Speciality.objects.get(id=target_id)
+        speciality.delete()
+        return speciality is None
 
 
 def update_speciality(name: str, code: str, level: str, form: str):
